Fix for numerical bug
In some cases, the initialization of some key variables (e.g. xl) may fail because the condition "zk(i) >= (z_zero-zini_depth)" is never met (in my case, it was because of numbers after the 6th decimal place). The fix ensures that, at the last iteration of the loop, the variables get initialized.