Correction to function Advection
If there is an overflow, the second timestep (ti=2) should still be executed! Otherwise there is less change in the input variables, or not at all if the water level was already at the maximum at the beginning. This can potentially make a large difference when there is inflow but the outflow is zero.